pyproject.toml
[tool.poetry]
name = "batch_data_pipeline"
version = "0.1.0"
description = "Batch data ingestion and processing pipeline."
authors = ["Teedot19 <taiwo_dotun@yahoo.com>"]
readme = "README.md"
packages = [
    { include = "batch_data_pipeline", from = "src" },
    { include = "orchestration"}
]

[tool.poetry.dependencies]
python = ">=3.10,<3.13"
dbt-core = "^1.9.0"  # ← Changed from "dbt"
pandas = "^2.2.3"
dagster = "^1.9.8"
dagster-webserver = "^1.9.8"
dbt-snowflake = "^1.9.2"
python-dotenv = "^1.0.1"
faker = "^38.0.0"
faker-commerce = "^1.0.4"
google-cloud-storage = "^3.5.0"
dagster-dbt = "^0.28.2"
pydantic = "^2.12.4"
email-validator = "^2.3.0"
rich = "^14.2.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.4"
pytest-mock = "^3.15.1"
ruff = "^0.14.5"
mypy = "^1.18.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.dagster]
module_name = "orchestration.definitions"